% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement.R
\name{agranaly}
\alias{agranaly}
\title{Interrater Agreement Analyses}
\usage{
agranaly(x, y, lev = unique(unlist(x)))
}
\arguments{
\item{x, y}{vectors of scores from two different raters for the same units
of observation. Alternatively, if \code{y} is missing, \code{x} can be a
matrix of scores with two columns, one per rater.}

\item{lev}{vector of all unique possible scores.}
}
\value{
Returns a list of agreement indices.
}
\description{
Functions for estimating agreement in scores from a pair of raters.
}

