% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critstudy.R
\name{critstudy}
\alias{critstudy}
\alias{crit_r}
\title{Criterion Validity Study}
\usage{
critstudy(x, y, covariates, rx = 1, ry = 1, ...)

crit_r(x, y = NULL, rx, ry)
}
\arguments{
\item{x, y}{vectors of scores of the same length from a test
\code{x} and the criterion it is intended to predict \code{y}.}

\item{covariates}{\code{data.frame} of covariates to control for when
finding the consequential validity ratio.}

\item{rx, ry}{reliabilities of \code{x} and \code{y} used to correct for
attenuation in the validity coefficient.}

\item{...}{further arguments passed to or from other functions.}
}
\value{
\code{critstudy} returns a list of results.
}
\description{
Functions for estimating validity coefficients
}
\examples{

ritems <- c("r414q02", "r414q11", "r414q06", "r414q09",
  "r452q03", "r452q04", "r452q06", "r452q07", "r458q01",
  "r458q07", "r458q04") |> paste0("s")
sitems <- c("s428q01", "s131q02d", "s131q04d", "s514q02", "s514q04",
  "s428q05", "s438q01t", "s514q03", "s465q01", "s465q02", "s438q02",
  "s465q04", "s438q03d", "s428q03", "s415q02", "s415q08t",
  "s415q07t")|> paste0("s")
pisa <- subset(PISA09, bookid == 3)
critstudy(x = rowSums(pisa[, ritems], na.rm = TRUE),
   y = rowSums(pisa[, sitems], na.rm = TRUE),
   covariates = pisa[, c("escs", "sex", "homelang", "pared")])

}
