% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{desanaly}
\alias{desanaly}
\title{Descriptive Analyses}
\usage{
desanaly(x, complete = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{vector, matrix, or data.frame of scores with one column
per variable.}

\item{complete}{logical with default \code{TRUE} indicating whether or not
\code{x} should be reduced to rows with complete data across all columns.}

\item{na.rm}{logical with default \code{FALSE} specifying whether missings
should be removed before calculating individual descriptives.}
}
\value{
Returns a matrix of descriptive statistics, including the mean,
median, standard deviation, skewness, kurtosis, minimum, maximum,
and number of complete cases (in columns) by variable (in rows).
}
\description{
Functions for obtaining descriptive statistics for a vector or matrix
of quantitative variables.
}

