% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difstudy.R
\name{difstudy}
\alias{difstudy}
\alias{mhd}
\title{Differential Item Functioning Study}
\usage{
difstudy(x, subset = 1:ncol(x), scores, complete = TRUE, na.rm = FALSE)

mhd(x, groups, scores)
}
\arguments{
\item{x}{matrix or data.frame of scored item responses.}

\item{subset}{optional vector for selecting a subset of columns from \code{x}.}

\item{scores}{optional vector of construct scores.}

\item{complete}{logical with default \code{TRUE} indicating whether or not
\code{x} should be reduced to rows with complete data across all columns.}

\item{na.rm}{logical with default \code{FALSE} specifying whether missings
should be removed before calculating individual descriptives.}

\item{groups}{optional grouping variable.}
}
\description{
Functions for estimating differential item functioning with a set of
scored item responses.
}

