% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstudy.R
\name{dstudy}
\alias{dstudy}
\alias{kurt}
\alias{skew}
\alias{sumcomp}
\alias{summiss}
\title{Descriptive Study}
\usage{
dstudy(x, complete = TRUE, na.rm = FALSE, ...)

kurt(x, na.rm = FALSE)

skew(x, na.rm = FALSE)

summiss(x, ...)

sumcomp(x, ...)
}
\arguments{
\item{x}{vector, matrix, or data.frame of scores with one column
per variable.}

\item{complete}{logical with default \code{TRUE} indicating whether or not
\code{x} should be reduced to rows with complete data across all columns.}

\item{na.rm}{logical with default \code{FALSE} specifying whether missings
should be removed before calculating individual descriptives.}
}
\value{
\code{dstudy} returns a data frame of descriptive statistics, including
the mean, median, standard deviation, skewness, kurtosis, minimum, maximum,
number of complete cases, and number of NAs (in columns), by variable
(in rows). \code{kurt} and \code{skew} return the kurtosis and skewness.
\code{summiss} and \code{sumcomp} count missings and complete cases.
}
\description{
Functions for obtaining descriptive statistics for a vector or matrix
of quantitative variables.
}
\examples{
dstudy(PISA09[, c("elab", "cstrat", "memor")])

}

