% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epmr-package.R
\docType{package}
\name{epmr-package}
\alias{epmr}
\alias{epmr-package}
\title{epmr: Tools for Educational and Psychological Measurement}
\description{
The epmr package contains tools for creating and evaluating tests in the social sciences, primarily in the contexts of education and psychology. The package accompanies the book Introduction to Educational and Psychological Measurement in R (free online, see link below). Example data come from the 2009 PISA study and the BDI-II. Functions currently support classical item and distractor analysis, reliability analysis, factor analysis, some item response theory, and analysis of differential item functioning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/talbano/epmr}
  \item \url{https://www.thetaminusb.com/intro-measurement-r/}
}

}
\author{
\strong{Maintainer}: Anthony Albano \email{tony.d.albano@gmail.com}

}
\keyword{internal}
