% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastudy.R
\name{fastudy}
\alias{fastudy}
\alias{plot.fastudy}
\title{Factor Analysis Study}
\usage{
fastudy(x, factors, complete = TRUE, ...)

\method{plot}{fastudy}(x, ylim = c(1, nrow(x$loadings)), ...)
}
\arguments{
\item{x}{matrix or data.frame of scored item responses, one row per person,
one column per item.}

\item{complete}{logical with default \code{FALSE} indicating whether or not
\code{x} should be reduced to rows with complete data across all columns.}

\item{...}{further arguments passed to or from other functions.}
}
\description{
Functions for examining the dimensionality of a data set via
exploratory factor analysis.
}

