% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstudy.R
\name{gstudy}
\alias{gstudy}
\alias{gstudy.default}
\alias{gstudy.formula}
\alias{gstudy.data.frame}
\alias{gstudy.merMod}
\title{Generalizability and Decision Studies}
\usage{
gstudy(x, ...)

\method{gstudy}{default}(x, data = NULL, ...)

\method{gstudy}{formula}(x, data = NULL, ...)

\method{gstudy}{data.frame}(x, random = TRUE, ...)

\method{gstudy}{merMod}(x, n, id = "person", ...)
}
\arguments{
\item{x}{for the default method, a \code{formula} object sent to the
\code{\link{lme4}} package. Otherwise, an object of class \code{\link{merMod}},
output by the \code{\link{lmer}} function. \code{x} can also be a data frame,
where a model with a single facet is assumed, with people in rows and the
facet in columns.}

\item{...}{further arguments passed to or from other methods.}

\item{data}{a data frame containing the variables named in formula. Required
when x is not a data frame.}

\item{random}{logical indicating whether the facet given in columns when
\code{x} is data frame should be treated as a random (default) or fixed
effect.}

\item{n}{vector of counts used in adjusting the g coefficient. Not currently
used.}

\item{id}{character string naming the id variable for the unit of observation,
defaulting to \code{"person"}.}
}
\value{
Returns a list containing the original call, model formula,
reliabilities, variance components, and n counts by facet.
}
\description{
Functions for running generalizability (G) and decision (D) studies
for estimating reliability in multi-faceted designs.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Method defaults to formula interface.

\item \code{formula}: Method for formula objects.

\item \code{data.frame}: Method for wide data frames with people as rows and a
single facet as columns

\item \code{merMod}: Method for objects of class \code{\link{merMod}}.
}}

