% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtstudy.R
\name{irtstudy}
\alias{irtstudy}
\alias{rirf}
\alias{riif}
\alias{rief}
\alias{rtrf}
\alias{rtif}
\alias{rtef}
\title{Item Response Theory Study}
\usage{
irtstudy(x, complete = FALSE, ...)

rirf(ip, theta = seq(-4, 4, length = 100))

riif(ip, theta = seq(-4, 4, length = 100))

rief(ip, theta = seq(-4, 4, length = 100))

rtrf(ip, theta = seq(-4, 4, length = 100))

rtif(ip, theta = seq(-4, 4, length = 100))

rtef(ip, theta = seq(-4, 4, length = 100))
}
\arguments{
\item{x}{matrix or data.frame of scored item responses, one row per person,
one column per item.}

\item{complete}{logical with default \code{FALSE} indicating whether or not
\code{x} should be reduced to rows with complete data across all columns.}

\item{...}{further arguments passed to or from other functions, including
\code{\link{glmer}}}

\item{ip}{vector or matrix of item parameters, with a in the first column, b
in the second, and c in the third.}

\item{theta}{vector of thetas over which to estimate the corresponding function.}
}
\description{
Functions for estimating item response theory item and person parameters.
Currently, only the 1 parameter Rasch model is supported.
}
