% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istudy.R
\name{istudy}
\alias{istudy}
\title{Scored Item Study}
\usage{
istudy(x, subset = 1:ncol(x), scores)
}
\arguments{
\item{x}{matrix or data.frame of scored item responses.}

\item{subset}{optional vector for selecting a subset of columns from \code{x}.}

\item{scores}{optional vector of construct scores used to calculate
item discrimination.}
}
\value{
Returns a list with three elements: \code{item} containing a
data.frame of item analysis output, \code{scale} containing descriptives
for the full scale, and \code{reliability} containing a vector of
internal consistency reliability estimates.
}
\description{
Functions for estimating classical item analysis indices, including
item difficulty (means and standard deviations by item),
item discrimination (item-total and corrected item-total correlations),
and alpha if item deleted. Descriptive statistics and reliability
estimates are also returned at the scale level.
}

