% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstudy.R
\name{mstudy}
\alias{mstudy}
\title{Measurement Scale Study}
\usage{
mstudy(x, subset = 1:ncol(x), scores, complete = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{matrix or data.frame of scored item responses.}

\item{subset}{optional vector for selecting a subset of columns from \code{x}.}

\item{scores}{optional vector of construct scores.}

\item{complete}{logical with default \code{TRUE} indicating whether or not
\code{x} should be reduced to rows with complete data across all columns.}

\item{na.rm}{logical with default \code{FALSE} specifying whether missings
should be removed before calculating individual descriptives.}
}
\value{
Returns a list with elements pertaining to each study, including:
\code{descriptives}, \code{items}, \code{reliability}, and \code{irt}.
}
\description{
Functions for running multiple studies on a set of scored item responses,
including a descriptive study of the total scale scores, and item,
internal consistency reliability, and item response theory studies.
}
