% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ostudy.R
\name{ostudy}
\alias{ostudy}
\title{Item Option Study}
\usage{
ostudy(x, groups, scores, cuts = c(0, 1/3, 2/3, 1), key, labels = c("low",
  "mid", "high"), itemid = colnames(x), filename, sep = ",")
}
\arguments{
\item{x}{matrix or data.frame of unscored item responses.}

\item{groups}{optional grouping variable as a vector with length
\code{nrow(x)}. When \code{cuts} are given, groups are created by comparing
\code{scores} to \code{cuts}.}

\item{scores}{optional vector of scores defining the ability or trait
levels of respondents. When \code{key} is provided, scores are calculated
by first scoring item responses in \code{x}.}

\item{cuts}{vector of cut scores used to define groups.}

\item{key}{vector of keyed responses, one per column in \code{x}.}

\item{labels}{character vector for labeling groups in output, with length
\code{length(groups)}.}

\item{itemid}{vector of labels for items, defaulting to column names in
\code{x}, if present, otherwise "Item1", "Item2", etc.}

\item{filename}{optional path to file where output will be written.}

\item{sep}{character string used to separate output in written file.}
}
\value{
Returns three lists of distractor analysis tables, one per item,
all stored within a larger list. These lists are \code{counts} for raw
frequencies, \code{rowpct} for percentages by row, and \code{colpct} for
percentages by column. Each table within each of these lists contains the
distribution of unscored item responses by group.
}
\description{
Functions for analyzing distributions of unscored item responses by
ability, also known as distractor analysis.
}

