% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoding.r
\name{recode}
\alias{recode}
\title{Recoding Data}
\usage{
recode(x, old, new = names(old), rev = TRUE)
}
\arguments{
\item{x}{vector of scores to be recoded}

\item{old}{list of old values in \code{x} which will be recoded according to
the name of the list element, unless new is specified.}

\item{new}{vector of new values to be assigned to corresponding value(s)
in \code{old}, assuming \code{old} and \code{new} are the same length.}

\item{rev}{logical, with default \code{FALSE}. If TRUE, old and new are
ignored and \code{x} is simply reverse coded. Assumes all possible values
are at least ordinal scale quantitative and observed.}
}
\value{
Returns a vector of recoded scores.
}
\description{
Recodes or reverse codes a vector of scores
}

