% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{relanaly}
\alias{alpha}
\alias{omega}
\alias{relanaly}
\title{Internal Consistency Reliability Analyses}
\usage{
relanaly(x, sigma = FALSE)

alpha(x, sigma = FALSE)

omega(x, sigma = FALSE)
}
\arguments{
\item{x}{a matrix of scored item responses, or a variance/covariance matrix
for a set of scored item responses when \code{sigma = TRUE}.}

\item{sigma}{boolean with default \code{FALSE} indicating whether or not
\code{x} is a variance/covariance matrix.}
}
\value{
Returns a vector of estimated reliability coefficient,
currently including alpha and omega.
}
\description{
Functions for estimating internal consistency reliability for a matrix
of scored item responses or a variance/covariance matrix.
}

