% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaling.R
\name{rescaling}
\alias{rescaling}
\alias{setmax}
\alias{setmean}
\alias{setmin}
\alias{setrange}
\alias{setsd}
\title{Rescaling Scores}
\usage{
setrange(x, y, to = range(y))

setmin(x, y, to = min(y))

setmax(x, y, to = max(y))

setmean(x, y, to = mean(y))

setsd(x, y, to = sd(y))
}
\arguments{
\item{x}{vector of scores to be modified.}

\item{y}{optional vector of scores from which new parameters will be taken.}

\item{to}{numeric value(s) that parameters in \code{x} will be set to,
defaulting to the \code{min}, \code{max}, \code{range}, \code{mean}, or
\code{sd} of \code{y}.}
}
\value{
Returns a vector of \code{length(x)} scores, with new parameters.
}
\description{
\code{setrange} truncates a quantitative variable to have a set minimum
and maximum. \code{setmin} truncates only to the minimum, and \code{setmax}
truncates only to the maximum. \code{setmean} and \code{setsd} set the
mean and standard deviation.
}

