% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsim.R
\name{rsim}
\alias{rsim}
\title{Simulating Correlated Scores}
\usage{
rsim(n, rho, x, meanx = 0, sdx = 1, meany = 0, sdy = 1)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
is taken to be the number required.}

\item{rho}{correlation in the population.}

\item{x}{vector of scores to be correlated with. If \code{x} is supplied,
\code{n} is overridden to be \code{length(x)}, meanx and sdx are set to
\code{mean(x)} and \code{sd(x)}, and y is created to
correlate at \code{rho} with \code{x}.}

\item{meanx, sdx}{the desired mean and standard deviation of x.}

\item{meany, sdy}{the desired mean and standard deviation of y.}
}
\value{
Returns a list containing:
\describe{
\item{x}{\code{x} when supplied, or the simulated vector of scores}
\item{y}{the vector y sampled from a population correlating with
\code{x} at \code{rho}}
\item{rho}{the supplied correlation \code{rho}}
\item{r}{the estimated correlation between x and y}
\item{meanx, sdx, meany, sdy}{the means and standard deviations of x and y}
}
}
\description{
Simulates vectors of scores with specified correlation, means, and
standard deviations.
}

