% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epmr.r
\docType{package}
\name{epmr}
\alias{epmr}
\alias{epmr-package}
\title{Tools for Educational and Psychological Measurement}
\description{
The epmr package contains tools for creating and evaluating tests in the
social sciences, primarily in the contexts of education and psychology.
The package accompanies the book Introduction to Educational and
Psychological Measurement (Albano, 2016).
}
\section{Functions}{

Functions currently support item, distractor, and reliability analyses.
}
\references{
Albano, A. D. (2016) Introduction to Educational and
Psychological Measurement: in R. Available at
\url{https://github.com/talbano/intro-measurement}
}

