% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastudy.R
\name{fastudy}
\alias{fastudy}
\alias{plot.fastudy}
\title{Factor Analysis Study}
\usage{
fastudy(x, factors, covmat, complete = TRUE, ...)

\method{plot}{fastudy}(x, ylim, h = 1, ...)
}
\arguments{
\item{x}{matrix or data.frame of scored item responses, one row per person,
one column per item.}

\item{factors}{the number of factors to extract.}

\item{covmat}{optional covariance matrix. Not used if \code{x} is supplied.}

\item{complete}{logical with default \code{FALSE} indicating whether or not
\code{x} should be reduced to rows with complete data across all columns.}

\item{...}{further arguments passed to or from other functions.}

\item{ylim}{vector of limits for y-axis in the scree plot, passed to \code{par}.}

\item{h}{y-axis value specifying a horizonal line the scree plot, passed to
\code{abline}.}
}
\value{
\code{fastudy} runs an exploratory factor analysis using the
\code{factanal} function, and returns a matrix of factor loadings. The
print method displays eigenvalues and proportion of variance explained
per factor.
}
\description{
Functions for examining the dimensionality of a data set via
exploratory factor analysis. \code{fastudy} is currently a simple
wrapper for factanal. \code{plot.fastudy} generates a scree plot
using \code{fastudy} output.
}
\examples{

# Exploratory factor analysis of the PISA approaches to learning scale
# Create vector of item names
items <- c("st27q01", "st27q03", "st27q05", "st27q07", "st27q04",
  "st27q08", "st27q10", "st27q12", "st27q02", "st27q06",
  "st27q09", "st27q11", "st27q13")

# Reduce PISA09 to complete data for Great Britain
pisa_gbr <- na.omit(PISA09[PISA09$cnt == "GBR", items])

# Fit EFA with six factors
fa_al <- fastudy(pisa_gbr, factors = 6)

# Scree plot
plot(fa_al, ylim = c(0, 2))

}
