% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstudy.R
\name{gstudy}
\alias{gstudy}
\alias{gstudy.default}
\alias{gstudy.formula}
\alias{gstudy.data.frame}
\alias{gstudy.merMod}
\title{Generalizability and Decision Studies}
\usage{
gstudy(x, ...)

\method{gstudy}{default}(x, data = NULL, ...)

\method{gstudy}{formula}(x, data = NULL, id = "person", ...)

\method{gstudy}{data.frame}(x, random = TRUE, column_label = "rater", id = "person", ...)

\method{gstudy}{merMod}(x, n, id = "person", ...)
}
\arguments{
\item{x}{for the default method, a \code{formula} object sent to the
\code{\link{lme4}} package. Otherwise, an object of class \code{\link{merMod}},
output by the \code{\link{lmer}} function. \code{x} can also be a data frame,
where a model with a single facet is assumed, with people in rows and the
facet in columns.}

\item{...}{further arguments passed to or from other methods.}

\item{data}{a data frame containing the variables named in formula. Required
when x is not a data frame.}

\item{id}{character string naming the id variable for the unit of observation,
defaulting to \code{"person"}.}

\item{random}{logical indicating whether the facet given in columns when
\code{x} is data frame should be treated as a random (default) or fixed
effect.}

\item{column_label}{string for labeling the column facet when \code{x} is
data frame.}

\item{n}{vector of counts used in adjusting the g coefficient. Not currently
used.}
}
\value{
Returns a list containing the original call, model formula,
reliabilities, variance components, and n counts by facet.
}
\description{
Functions for running generalizability (G) and decision (D) studies
for estimating reliability in multi-faceted designs.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Method defaults to formula interface.

\item \code{formula}: Method for formula objects.

\item \code{data.frame}: Method for wide data frames with people as rows and a
single facet as columns

\item \code{merMod}: Method for objects of class \code{\link{merMod}}.
}}

\examples{
# Reading items
ritems <- c("r414q02", "r414q11", "r414q06", "r414q09",
  "r452q03", "r452q04", "r452q06", "r452q07", "r458q01",
  "r458q07", "r458q04")
rsitems <- paste0(ritems, "s")

# Subset of PISA09 data for Belgium
pisa <- subset(PISA09, cnt == "BEL")[, rsitems]

# Convert to long format
np <- nrow(pisa)
ni <- ncol(pisa)
pisal <- data.frame(person = factor(rep(1:np, ni)),
  item = rep(rsitems, each = np),
  score = unlist(pisa), row.names = 1:(np * ni))

# Compare coefficient alpha using rstudy and gstudy
rout <- rstudy(pisa, use = "complete.obs")
gout <- gstudy(pisa, column_label = "item")
rout$alpha
gout$g

# Alternatively, find g using long data and formula method
# Results differ slightly because missing data are excluded with the
# data frame method but included by default with the formula method
goutl <- gstudy(score ~ (1 | person) + (1 | item), data = pisal)
goutl$g

}
