% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoding.r
\name{recode}
\alias{recode}
\title{Recoding Data}
\usage{
recode(x, from, to = names(from), reverse = missing(from),
  trynumeric = TRUE)
}
\arguments{
\item{x}{vector of scores to be recoded.}

\item{from}{list of old values in \code{x} which will be recoded according to
the name of the list element, unless \code{to} is specified.}

\item{to}{vector of new values to be assigned to corresponding value(s)
in \code{from}, assuming \code{from} and \code{to} are the same length.}

\item{reverse}{logical, with default \code{TRUE} when \code{from} is missing.
If TRUE, old and new are ignored and \code{x} is simply reverse coded.
Assumes all possible values are at least ordinal scale quantitative and
observed.}

\item{trynumeric}{logical with default \code{TRUE}. If \code{TRUE}, an
attempt is made to convert values in \code{to} to \dQuote{\code{numeric}}.
\code{as.numeric(to)} must first succeed without warnings.}
}
\value{
Returns a vector of recoded scores.
}
\description{
Recodes or reverse codes a vector of scores
}
